language added

+++ app/controllers/crudControllerAdmin copy.js
@@ -1,4 +1,4 @@
-///this controller for users
+///this controller for users 
 const asyncHandler = require("../middleware/async");
 const User = require("../models/User");
 const { where } = require("sequelize");
@@ -6,10 +6,12 @@ const { where } = require("sequelize");
 // @desc      Get  Baby get all as history
 // @route     GET /api/v1/route/history
 // @access    Private
-exports.getHistory = (Model, message) => {
+exports.getHistory = (Model) => {
   return asyncHandler(async (req, res, next) => {
-    if (res.advancedResults) return res.status(200).json(res.advancedResults);
-    const data = await Model.findAll();
+    // Check if the requesting mother owns the specified baby
+    const data = await Model.findAll({
+      where: { user_id: req.user.id },
+    });
     if (!data) {
       return res.status(403).json({
         success: false,
@@ -18,14 +20,14 @@ exports.getHistory = (Model, message) => {
     }
 
     // Get the feed history for the specified baby
-    res.status(200).json({ success: true, message, data });
+    res.status(200).json({ success: true, message: "Found history", data });
   });
 };
 
 // @desc      Get single
 // @route     GET /api/v1/route/:modelPk
 // @access    Private
-exports.getOne = (Model, message) => {
+exports.getOne = (Model) => {
   return asyncHandler(async (req, res, next) => {
     // Extract baby ID from the request params or body
     const { modelPk } = req.params;
@@ -34,28 +36,34 @@ exports.getOne = (Model, message) => {
     const data = await Model.findOne({
       where: {
         id: modelPk,
+        user_id: req.user.id,
       },
     });
 
-    res.status(200).json({ success: true, message, data });
+    res.status(200).json({ success: true, message:'Data found', data });
   });
 };
 
 // @desc      Create
 // @route     POST /api/v1/route/:babyId
 // @access    Private
-exports.create = (Model, message) => {
+exports.create = (Model) => {
   return asyncHandler(async (req, res, next) => {
+    // Extract baby ID from the request params or body
+    const { babyId } = req.params;
+
+    // Get the feed history for the specified baby
+    req.body.user_id = req.user.id;
     const data = await Model.create(req.body);
 
-    res.status(200).json({ success: true, message, data });
+    res.status(200).json({ success: true, message: "created", data });
   });
 };
 
 // @desc      Update
 // @route     PUT /api/v1/babyfeed/:babyId/:modelPk
 // @access    Private
-exports.update = (Model, message) => {
+exports.update = (Model) => {
   return asyncHandler(async (req, res) => {
     // Extract baby ID from the request params or body
     const { modelPk } = req.params;
@@ -64,33 +72,34 @@ exports.update = (Model, message) => {
     const updated = await Model.update(req.body, {
       where: {
         id: modelPk,
+        user_id: req.user.id,
       },
     });
 
     if (!updated[0]) {
       return res
         .status(200)
-        .json({ success: false, message: "Record not modified" });
+        .json({ success: false, message: "Record no modified" });
     }
 
-    res.status(200).json({ success: true, message });
+    res.status(200).json({ success: true, message: "updated" });
   });
 };
 
 // @desc      Delete
 // @route     DELETE /api/v1/babyfeed/:babyId/:modelPk
 // @access    Private/Admin
-exports.deleteOne = (Model, message) => {
+exports.deleteOne = (Model) => {
   return asyncHandler(async (req, res) => {
     // Extract baby ID from the request params or body
     const { modelPk } = req.params;
 
     // Get the feed history for the specified baby
     const deleted = await Model.destroy({
-      where: { id: modelPk },
+      where: { id: modelPk, user_id: req.user.id },
     });
 
-    res.status(200).json({ success: true, message });
+    res.status(200).json({ success: true, message: "deleted" });
   });
 };
 
@@ -100,10 +109,7 @@ exports.deleteOne = (Model, message) => {
 exports.deleteAll = (Model) => {
   return asyncHandler(async (req, res) => {
     // Get the feed history for the specified baby
-    const deleted = await Model.destroy({
-      where: {},
-      truncate: false,
-    });
+    const deleted = await Model.destroy({ where: { user_id: req.user.id } });
 
     res.status(200).json({ success: true, message: "All deleted" });
   });
